(Feat) The shared buffer is implmented for ONV.
The shared buffer which first has been implmented for kernel modules,
transformed to a user-space code compatibale with DPDK. The host
specification functionality has been removed for now. Besides that,
instead of shared kernel variables, it works based shared memory. The
metadate circular buffer is removed and now the metadata of each cell
can be considered as a part of the cell itself.